% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_plot_area.R
\name{plot_tree_plot_area}
\alias{plot_tree_plot_area}
\title{Plot Tree Plot Area}
\usage{
plot_tree_plot_area(
  angle_counts,
  boundaries,
  tnr,
  enr,
  bnr = NULL,
  frame_factor = 1,
  use_sub = NULL,
  counting_factor = 4
)
}
\arguments{
\item{angle_counts}{A \code{\link{data.frame}} containing angle counts.
It has to have columns named by the contents of
either\cr
\code{\link{get_defaults}()[["angle_counts"]]} or \cr
\code{getOption("treePlotArea")[["angle_counts"]]}.\cr
Could be
\code{bw2bwi2022de(get(data("trees", package = "treePlotArea")))}).}

\item{boundaries}{A \code{\link{data.frame}} containing boundaries.
It has to have columns named by the contents of
either\cr
\code{\link{get_defaults}()[["boundaries"]]} or \cr
\code{getOption("treePlotArea")[["boundaries"]]}.\cr
Could be
\code{get(data("boundaries", package = "treePlotArea"))} or the
output of
\code{\link{get_boundary_polygons}}.}

\item{tnr}{Number of the tract.}

\item{enr}{Number of the tract's corner.}

\item{bnr}{If given, the number of a corner's tree.}

\item{frame_factor}{Plotting from as a factor of the tree plot area. Stick
with the default.}

\item{use_sub}{Deprecated.}

\item{counting_factor}{The basal area factor used in counting the trees. For
tally trees in the German national forest inventory its value is 4 [m^2].}
}
\value{
A list with the corrections factors for the plot's trees' plot areas,
the plot's trees
and the plot's boundaries.
}
\description{
Visualize a corner, its boundaries and tree plot areas.
}
\examples{
tnr <- 166
enr <- 2
bnr <- 7
angle_counts <- bw2bwi2022de(get(data("trees", package = "treePlotArea")))
boundaries <- get(data("boundaries", package = "treePlotArea"))
plot_tree_plot_area(angle_counts = angle_counts,
                    boundaries = boundaries,
                    tnr =  tnr, enr = enr, bnr = bnr, frame_factor = 4)
plot_tree_plot_area(angle_counts = angle_counts,
                    boundaries = boundaries,
                    tnr =  tnr, enr = enr, frame_factor = 1)
tnr <- 9823
enr <- 1
bnr <- 2
plot_tree_plot_area(angle_counts = angle_counts,
                    boundaries = boundaries,
                    tnr =  tnr, enr = enr, bnr = bnr, frame_factor = 4)
plot_tree_plot_area(angle_counts = angle_counts,
                    boundaries = boundaries,
                    tnr =  tnr, enr = enr, frame_factor = 1)
}
