% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2equation.R
\name{points2equation}
\alias{points2equation}
\title{Get a Slope–intercept Form from a Two-point Form of an Equation}
\usage{
points2equation(p1, p2 = c(0, 0))
}
\arguments{
\item{p1}{The first point (x, y).}

\item{p2}{The second point (x, y).}
}
\value{
A named vector with intercept ["a"] and slope ["b"].
If both points have the same value for x, no function exists. Then
the intercept is \code{\link{NA}} and the slope gives the value of x.
}
\description{
Two-point from is often seen in German national forest inventory data,
we want to get an equation of form
\emph{y = a + bx}.
}
\examples{
points2equation(c(0, 4), c(1, 5))
}
\keyword{internal}
