% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlotArea-package.R
\docType{package}
\name{treePlotArea-package}
\alias{treePlotArea}
\alias{treePlotArea-package}
\title{Correction Factors for Tree Plot Areas Intersected by Stand
Boundaries}
\description{
The German national forest inventory uses angle count sampling, a
sampling method first published by Bitterlich (1947) and extended by
Grosenbaugh (1952) as probability proportional to size sampling. When plots
are located near stand boundaries, their sizes and hence their
probabilities need to be corrected.
}
\details{
You will find the details in\cr
\code{vignette("An_Introduction_to_treePlotArea", package =
"treePlotArea")}.
}
\references{
\cite{
Bitterlich, W. (1947): Die Winkelzählmessung.
Allgemeine Forst- und Holzwirtschaftliche Zeitung, 58.
}

\cite{
Grosenbaugh, L. R. (1952): Plotless Timber Estimates -- New, Fast, Easy.
Journal of Forestry.
https://academic.oup.com/jof/article-abstract/50/1/32/4684174.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/fvafrcu/treeplotarea.git}
}

}
\author{
\strong{Maintainer}: Andreas Dominik Cullmann \email{fvafrcu@mailbox.org}

Other contributors:
\itemize{
  \item Bernhard Bösch [contributor]
  \item Christoph Fischer [contributor]
  \item Gerald Kändler [contributor]
}

}
\keyword{package}
