% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_length.R
\name{vector_length}
\alias{vector_length}
\title{Give the Length of a Vector}
\usage{
vector_length(p1, p2 = c(0, 0))
}
\arguments{
\item{p1}{The first point (x, y).}

\item{p2}{The second point (x, y).}
}
\value{
The length of the vector.
}
\description{
A vector given by two points in a plane.
}
\examples{
p1 <- c(0, 4)
vector_length(p1)
p2 <- c(3, 4)
vector_length(p1, p2)
}
\seealso{
Other geometry functions: 
\code{\link{get_intersection}()},
\code{\link{points2equation}()}
}
\concept{geometry functions}
\keyword{internal}
